Se agregaron nuevos estados a la aplicación, asimimsmo se agregó la lógica necesaria para las visitas incluyendo la parte de los correos de notificación

+++ src/main/java/com/espiritware/opusclick/service/CityServiceImpl.java
@@ -14,10 +14,10 @@ import com.espiritware.opusclick.model.City;
 public class CityServiceImpl implements CityService{
 	
 	
-	GenericDao< City,Long > cityDao;
+	GenericDao< City,Integer > cityDao;
 
 	@Autowired
-	public void setDao( GenericDao< City,Long > daoToSet ){
+	public void setDao( GenericDao< City,Integer> daoToSet ){
 		cityDao = daoToSet;
 		cityDao.setEntityClass( City.class );
 	}
